requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
django==5.0; python_version >= '3.10'
django-filter==23.5; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
et-xmlfile==1.1.0; python_version >= '3.6'
idna==3.6; python_version >= '3.5'
mysqlclient==2.2.0; python_version >= '3.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
urllib3==2.1.0; python_version >= '3.8'
